chore: split files in new storer
