+++ pkg/localstorev2/netstore.go
@@ -27,7 +27,7 @@ func (db *DB) DirectUpload() PutterSession {
 	// already encountered error.
 	eg, egCtx := errgroup.WithContext(context.Background())
 
-	return &putterSessionImpl{
+	return &putterSession{
 		Putter: storage.PutterFunc(func(ctx context.Context, ch swarm.Chunk) error {
 			workers <- struct{}{}
 			eg.Go(func() error {
